Второй вариант

+++ Task010/Program.cs
@@ -15,5 +15,5 @@ Console.WriteLine("Введите число: ");
 int N = int.Parse(Console.ReadLine());
 if (N < 0) N = -N;
 while(N > 999) N /=10;
-long number = N % 10;
+int number = N % 10;
 Console.WriteLine("Третья цифра = " + number);
